+++ test/test_parse_value.h
@@ -693,7 +693,7 @@ UFBXT_TEST(parse_array_bad_conversion)
 	ufbxt_assert(ufbxi_parse_value(uc, 'b', &arr));
 	ufbxt_assert(!ufbxi_decode_array(uc, &arr, data));
 
-	uint64_t begin = uc->focused_node.next_value_pos;
+	size_t begin = uc->focused_node.next_value_pos;
 	ufbxt_assert(ufbxi_parse_value(uc, 'f', &arr));
 	ufbxt_assert(!ufbxi_decode_array(uc, &arr, data));
 	uc->focused_node.next_value_pos = begin;
